Add files via upload
Current ones not bad. 90% accuracy. uses a training set of ~58k and a test set of ~5.8k. Will improve later but maybe use this to start statistical analysis. The tree itself is about 180MB but if you reduce ntree to 100 its about 40ish. and it looks like the accuracy is not drastically improved when using 500 instead of 100.